Create ForLoop_CPP
For each integer i in the interval [a,b] :

If ,1 <= i <= 9 then print the English representation of it in lowercase. That is "one" for 1, "two" for 2, and so on.
Else if i > 9 and it is an even number, then print "even".
Else if i > 9 and it is an odd number, then print "odd".