Create ChocalateFest_C
The first line contains an integer,t , denoting the number of test cases to analyze.
Each of the next t lines contains three space-separated integers: n,c and m. They represent money to spend, cost of a chocolate, and the number of wrappers he can turn in for a free chocolate.