Create Addition.py
Recently, Chef studied the binary numeral system and noticed that it is extremely simple to perform bitwise operations like AND, XOR or bit shift on non-negative integers, while it is much more complicated to perform arithmetic operations (e.g. addition, multiplication or division).

After playing with binary operations for a while, Chef invented an interesting algorithm for addition of two non-negative integers A and B:

function add(A, B):
    while B is greater than 0:
        U = A XOR B
        V = A AND B
        A = U
        B = V * 2
    return A
Now Chef is wondering how fast this algorithm is. Given the initial values of A and B (in binary representation), he needs you to help him compute the number of times the while-loop of the algorithm is repeated.