Update insertionsort.py
Time Complexities

Worst Case Complexity: O(n2)

Suppose, an array is in ascending order, and you want to sort it in descending order. In this case, worse case complexity occers.

Each element has to be compared with each of the other elements so, for every nth element, (n-1) number of comparisons are made.

Thus, the total number of comparisons = n*(n-1) ~ n2
 
Best Case Complexity: O(n)

When the array is already sorted, the outer loop runs for n number of times whereas the inner loop does not run at all. So, there is only n number of comparison. Thus, complexity is linear.
 
Average Case Complexity: O(n2)

It occurs when the elements of a array are in jumbled order (neither ascending nor descending).
Space Complexity

Space complexity is O(1) because an extra variable key is used.

Insertion Sort Applications
The insertion sort is used when:

the array is has a small number of elements
there are only a few elements left to be sorted