Create For_The_Benefit_Of_All.c
The only difference between easy and hard versions is the constraints.

In the new secret NASA, there are n astronauts. Professor Brand wants to form a team headed by Dr Mann for the Lazarus mission. Astronauts will team up in pairs for the mission. But there is one condition: two astronauts will team up with one another if and only if each person believes that the other person is good enough for him.

An astronaut considers another good enough for him if and only if the other person has an astronaut rating of at least half of his own.

The professor now wonders, how many pairs are possible among the astronauts. Find the number of pairs possible.