Create Strange_Game.py
Alice and Bob are very good friends. They keep on playing some strange games. Today Alice came up with a new game, in which each player gets n cards at the start. Each card has it's value printed on it.
The game proceeds as each player chooses a random card and shows it's value. The player having card with higher value wins. As Alice came up with this game, he wants to ensure his win. So he starts to increase value of some cards using an algorithm. To increase the value of a card by 1, the running time of algorithm is k seconds.
Find the minimum running time of algorithm, ensuring the win of Alice.