Create DJ_Party.py
.During the DJ party, gentleman and Ladies, formed a queue of n people in the bar. Initially the all stood in the order they entered the bar. However, after a while the gentleman started feeling awkward for standing in front of the ladies in the queue and they started letting the ladies move forward each second.

Lets describe the process more precisely. Let's say that the positions in the queue are sequentially numbered by integers from 1 to n, at that the person in the position number 1 is served first. Then, if at time x a gentleman stands on the i-th position and a lady stands on the (i + 1)-th position, then at time x + 1 the i-th position will have a lady and the (i + 1)-th position will have a gentleman. The time is given in seconds.

You've got the initial position of all people, at the initial moment of time. Determine the way the queue is going to look after t seconds.