Create Jolites.c
All submissions for this problem are available.Striver is the coolest boy of the college and everyone loves him too.

He was studying for his end semester exams. Meanwhile he started looking for his favourite string "JGEC" in every sentence. He is counting the number of times the word is written in a sentence.

A sentence in his college only consists of capital english alphabets without any spaces. Your task is to check his counting.