Create Pass_XOR_Fail.py
Pappu wants to attend the Machine Learning workshop at Pravega. However, the tickets are sold out, and the only way to get in is to get an X-clusive VIP Pass. To get that, Pappu needs to pass a test: he will be given a single non-negative integer R and he will have to reply with the XOR (exclusive or) of all the integers between 0 and R, including R.

Can you help Pappu pass the test?