Create Matryoshka_Dolls.py
The Cricket World Cup 2015 is nearing and it will be held in Australia and New Zealand. The organizing committee is planning to give a welcome gift to each team member. After careful thought and planning, they have decided on Matryoshka dolls. 

A Matryoshka doll refers to a set of wooden dolls of strictly decreasing size, placed one inside the other. Any doll can contain only one doll directly inside it.

Given the sizes of N dolls, output "YES" if it is possible to nest them all and have one doll on the outside and "NO" otherwise.