Create The_Saviours.cpp
Due to some calamity, all the roads near Banasthali University got damaged. The villages situated near-by began experiencing a shortage of food.

Since nobody else was able to help them immediately, Banasthali University took the initiative to supply them with sufficient food till some external help reached the village. Banasthali has it's own flying club which decided to drop food packages to the houses.

All the houses in the village are in a straight line. Our pilot, Drishti, was given the responsibility to drop the packages. Consider that all the houses have equal number of members and that one food package can feed two houses. To maintain decorum in the village while dropping the packages, a rule was made. According to this rule, only the next two houses from the position where the package is dropped can feed on that package. Note that, the basic motive is to feed all the houses.

Given the way the packages are dropped, can you tell if Drishti was able to distribute food for every house or not?